[Crystal] Temporarily fix multi-processing mode of Amber server (#1392)
The priority of the settings in `config/application.cr` does not seem to
be higher than the settings in `config/environments/*.yml`.
And the cluster mode of Amber is buggy at the moment. If the setting of
`process_count: 1` in `config/environments/*.yml` is changed to
`process_count: 4`, then the result is 8 sub-processes are forked.
The simplest workaroud at the moment is to fork `System.cpu_count`
sub-processes in `src/amber.cr`.