Use a single thread for the server
When using `nproc` X 16 threads, there are 16 times the number of concurrent threads then there are processor cores.

This slows down the machine due to excessive lock contention and context switches. It only makes sense if most of the time the threads are "sleeping" or waiting on external IO (database requests).

For this simple plaintext test, a single thread would be better.